change name style
